Add showMoves to Rook and Pawn classes
Rooks and Pawns can now list their moves. This needs refactoring:

* It doesn't take into account whether pieces are capturable
* There's a lot of repeated code

It would be good to be able to pass Position an algebraic expression and have
the position class (or a dedicated class) search for available moves. This
would save a lot of repeated code. Something like this:

new MoveBuilder()
    .origin(Position)
    .transformY("+1")
    .transformX("-1")
    .list()

This would look in the north-east diagonal line for positions to move - as in
the case of a bishop.

I'm still not sure though. TBC.