Add an initial MoveBuilder class to calculate moves
This MoveBuilder is an implementation of the idea put forward in the commit
message of 19cf1a3.

It allows us to claculate a list of moves for any piece by using a short,
less-repetitive syntax free of loops. This allows us to tidy our Piece classes
and means we don't need to write repetetive, error-prone loops and collision
detection for each piece.

It is implemented using the Builder pattern (see
http://en.wikipedia.org/wiki/Builder_pattern). Rows and colums are manipulated
for direction by a string which is tokenized during setting. For example, to
list all possible moves north for a Rook, transform the rows like so:

    .setTransformRow("+1);

This is tokenized into an Add operator with a value of 1. Before checking
a square to see if we can move, we take the piece's position and add 1 (+1) to
the row. If the move is allowed, we transform the row again and re-check until
the board is out of bounds or a piece is detected (and can't be captured).