Update Base piece class with a MoveBuilder helper
This helper lets us create a new MoveBuilder for the current piece, with the
piece and origin properties already set. A 3 liner becomes a 1 liner :)

We've also separated showMoves into two methods - findPossibleMoves(), which
lists all possible moves as an array of Positions, and showMoves() which
returns an array of strings as required by the "list" command.

Also fix indentations caused by IntelliJ (what the ****? this editor is
crazy!).